vision/fruit_not_fruit/requirements.txt
absl-py==0.4.0
astor==0.7.1
certifi==2018.8.13
cycler==0.10.0
gast==0.2.0
grpcio==1.14.1
h5py==2.8.0
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib==2.2.3
numpy==1.15.1
opencv-python==3.4.2.17
Pillow==5.2.0
protobuf==3.6.1
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
scipy==1.1.0
six==1.11.0
tensorboard==1.9.0
tensorflow==1.9.0
termcolor==1.1.0
Werkzeug==0.14.1
